+++ es/ae/contract.js
@@ -63,10 +63,12 @@ async function handleCallError (result) {
  * @param {String} source Contract source code
  * @param {String} name Name of function to call
  * @param {Array} args Argument's for call
+ * @param {Object} [options={}]  Options
+ * @param {Object} [options.filesystem={}] Contract external namespaces map
  * @return {Promise<String>}
  */
-async function contractEncodeCall (source, name, args) {
-  return this.contractEncodeCallDataAPI(source, name, args)
+async function contractEncodeCall (source, name, args, options) {
+  return this.contractEncodeCallDataAPI(source, name, args, options)
 }
 
 /**
@@ -78,10 +80,11 @@ async function contractEncodeCall (source, name, args) {
  * @param {String } fn - function name
  * @param {String} callValue - result call data
  * @param {String} callResult - result status
+ * @param {Object} [options={}]  Options
+ * @param {Object} [options.filesystem={}] Contract external namespaces map
  * @return {Promise<String>} Result object
  * @example
  * const decodedData = await client.contractDecodeData(SourceCode ,'functionName', 'cb_asdasdasd...', 'ok|revert')lt
- * @param options
  */
 async function contractDecodeData (source, fn, callValue, callResult, options) {
   return this.contractDecodeCallResultAPI(source, fn, callValue, callResult, options)
@@ -96,10 +99,11 @@ async function contractDecodeData (source, fn, callValue, callResult, options) {
  * @param {String} address Contract address
  * @param {String} name Name of function to call
  * @param {Array} args  Argument's for call function
- * @param {Object} options [options={}]  Options
- * @param {String} top [options.top] Block hash on which you want to call contract
+ * @param {Object} [options={}]  Options
+ * @param {String} [options.top] Block hash on which you want to call contract
  * @param bytecode
  * @param {String} options [options.options]  Transaction options (fee, ttl, gas, amount, deposit)
+ * @param {Object} filesystem [options.options.filesystem] Contract external namespaces map
  * @return {Promise<Object>} Result object
  * @example
  * const callResult = await client.contractCallStatic(source, address, fnName, args = [], { top, options = {} })
@@ -115,7 +119,7 @@ async function contractCallStatic (source, address, name, args = [], { top, opti
     : await this.address().catch(e => opt.dryRunAccount.pub)
 
   // Prepare call-data
-  const callData = await this.contractEncodeCall(source, name, args)
+  const callData = await this.contractEncodeCall(source, name, args, options)
 
   // Get block hash by height
   if (top && !isNaN(top)) {
@@ -176,8 +180,8 @@ async function dryRunContractTx (tx, callerId, source, name, opt = {}) {
  * @param {String} address Contract address
  * @param {String} name Name of function to call
  * @param {Array} args Argument's for call function
- * @param {Object} options Transaction options (fee, ttl, gas, amount, deposit)
- * @return {Promise<Object>} Result object
+ * @param {Object} [options={}] Transaction options (fee, ttl, gas, amount, deposit)
+ * @param {Object} [options.filesystem={}] Contract external namespaces map* @return {Promise<Object>} Result object
  * @example
  * const callResult = await client.contractCall(source, address, fnName, args = [], options)
  * {
@@ -192,7 +196,7 @@ async function contractCall (source, address, name, args = [], options = {}) {
   const tx = await this.contractCallTx(R.merge(opt, {
     callerId: await this.address(opt),
     contractId: address,
-    callData: await this.contractEncodeCall(source, name, args)
+    callData: await this.contractEncodeCall(source, name, args, opt)
   }))
 
   const { hash, rawTx } = await this.send(tx, opt)
@@ -203,7 +207,7 @@ async function contractCall (source, address, name, args = [], options = {}) {
       hash,
       rawTx,
       result,
-      decode: () => this.contractDecodeData(source, name, result.returnValue, result.returnType)
+      decode: () => this.contractDecodeData(source, name, result.returnValue, result.returnType, opt)
     }
   } else {
     await this.handleCallError(result)
@@ -218,7 +222,8 @@ async function contractCall (source, address, name, args = [], options = {}) {
  * @param {String} code Compiled contract
  * @param {String} source Contract source code
  * @param {Array} initState Arguments of contract constructor(init) function
- * @param {Object} options Transaction options (fee, ttl, gas, amount, deposit)
+ * @param {Object} [options={}] Transaction options (fee, ttl, gas, amount, deposit)
+ * @param {Object} [options.filesystem={}] Contract external namespaces map* @return {Promise<Object>} Result object
  * @return {Promise<Object>} Result object
  * @example
  * const deployed = await client.contractDeploy(bytecode, source, init = [], options)
@@ -234,7 +239,7 @@ async function contractCall (source, address, name, args = [], options = {}) {
  */
 async function contractDeploy (code, source, initState = [], options = {}) {
   const opt = R.merge(this.Ae.defaults, options)
-  const callData = await this.contractEncodeCall(source, 'init', initState)
+  const callData = await this.contractEncodeCall(source, 'init', initState, opt)
   const ownerId = await this.address(opt)
 
   const { tx, contractId } = await this.contractCreateTx(R.merge(opt, {
@@ -253,8 +258,8 @@ async function contractDeploy (code, source, initState = [], options = {}) {
       transaction: hash,
       rawTx,
       address: contractId,
-      call: async (name, args = [], options) => this.contractCall(source, contractId, name, args, R.merge(opt, options)),
-      callStatic: async (name, args = [], options = {}) => this.contractCallStatic(source, contractId, name, args, { ...options, options: { onAccount: opt.onAccount, ...options.options } }),
+      call: async (name, args = [], options = {}) => this.contractCall(source, contractId, name, args, R.merge(opt, options)),
+      callStatic: async (name, args = [], options = {}) => this.contractCallStatic(source, contractId, name, args, { ...options, options: { onAccount: opt.onAccount, ...R.merge(opt, options.options) } }),
       createdAt: new Date()
     })
   } else {
@@ -268,7 +273,8 @@ async function contractDeploy (code, source, initState = [], options = {}) {
  * @alias module:@aeternity/aepp-sdk/es/ae/contract
  * @category async
  * @param {String} source Contract sourece code
- * @param {Object} options Transaction options (fee, ttl, gas, amount, deposit)
+ * @param {Object} [options={}] Transaction options (fee, ttl, gas, amount, deposit)
+ * @param {Object} [options.filesystem={}] Contract external namespaces map* @return {Promise<Object>} Result object
  * @return {Promise<Object>} Result object
  * @example
  * const compiled = await client.contractCompile(SOURCE_CODE)
@@ -279,11 +285,12 @@ async function contractDeploy (code, source, initState = [], options = {}) {
  * }
  */
 async function contractCompile (source, options = {}) {
-  const bytecode = await this.compileContractAPI(source, options)
+  const opt = R.merge(this.Ae.defaults, options)
+  const bytecode = await this.compileContractAPI(source, opt)
   return Object.freeze(Object.assign({
-    encodeCall: async (name, args) => this.contractEncodeCall(source, name, args),
-    deploy: async (init, options = {}) => this.contractDeploy(bytecode, source, init, options),
-    deployStatic: async (init, options = {}) => this.contractCallStatic(source, null, 'init', init, { bytecode, top: options.top, options })
+    encodeCall: async (name, args) => this.contractEncodeCall(source, name, args, R.merge(opt, options)),
+    deploy: async (init, options = {}) => this.contractDeploy(bytecode, source, init, R.merge(opt, options)),
+    deployStatic: async (init, options = {}) => this.contractCallStatic(source, null, 'init', init, { bytecode, top: options.top, options: R.merge(opt, options) })
   }, { bytecode }))
 }